Implement a ShaderProgram handler
The ShaderProgram handler is currently capable of loading and compiling
a vertex and a fragment shader, but should be extendable to all kinds
of shaders. Then a shader program is created, which can be used in a
main method by invoking shaderProgram.use().

Both the shaders and the program are checked for compilation and linking
errors. If not successful, this should throw an exception.