Add uniform update to shader handler
Add method to update mat4 in shader handler. There are going to be more
methods like this for every kind of possible uniform value. #2

To assure that the uniform name exists in the shader a method to find the
location was implemented, which throws an exception if the location is -1.