Implement interop example for rendering
This example only works for windows right now, but since it is only a
draft anyway and will probably updated with opencl handlers in the future,
this does not really matter.

This example will be used to create the necessary structures for OpenCL
OpenGL interoperability. But besides that, it actually works pretty well.