Move VBOHandler into BufferHandler
Combine both particle buffer handlers into one, since they actually work
together for the particles.
The rationale is the necessary access to parts of vbos through index
arrays which have to be implemented still. These should also be
incorporated into this buffer handler. Although that might be subject to
change.